editoast: models: move type `Error` at crate root
