+++ editoast/editoast_derive/src/model/codegen/create_batch_with_key_impl.rs
@@ -49,14 +49,14 @@ impl ToTokens for CreateBatchWithKeyImpl {
                     .returning((#(dsl::#columns,)*))
                     .load_stream::<#row>(conn.write().await.deref_mut())
                     .await
-                    .map_err(|e| Self::Error::from(editoast_models::model::Error::from(e)))?
+                    .map_err(|e| Self::Error::from(editoast_models::Error::from(e)))?
                     .map_ok(|row| {
                         let model = <#model as Model>::from_row(row);
                         (model.get_id(), model)
                     })
                     .try_collect::<Vec<_>>()
                     .await
-                    .map_err(|e| Self::Error::from(editoast_models::model::Error::from(e)))?
+                    .map_err(|e| Self::Error::from(editoast_models::Error::from(e)))?
             },
         };
