+++ editoast/editoast_derive/src/model/codegen/delete_static_impl.rs
@@ -45,7 +45,7 @@ impl ToTokens for DeleteStaticImpl {
                         .execute(conn.write().await.deref_mut())
                         .await
                         .map(|n| n == 1)
-                        .map_err(|e| Self::Error::from(editoast_models::model::Error::from(e)))
+                        .map_err(|e| Self::Error::from(editoast_models::Error::from(e)))
                 }
             }
         });
