+++ editoast/editoast_derive/src/model/codegen/list_impl.rs
@@ -68,11 +68,11 @@ impl ToTokens for ListImpl {
                         .select((#(dsl::#columns,)*))
                         .load_stream::<#row>(conn.write().await.deref_mut())
                         .await
-                        .map_err(|e| Self::Error::from(editoast_models::model::Error::from(e)))?
+                        .map_err(|e| Self::Error::from(editoast_models::Error::from(e)))?
                         .map_ok(<#model as crate::models::prelude::Model>::from_row)
                         .try_collect::<Vec<_>>()
                         .await
-                        .map_err(|e| Self::Error::from(editoast_models::model::Error::from(e)))
+                        .map_err(|e| Self::Error::from(editoast_models::Error::from(e)))
                 }
             }
