+++ editoast/editoast_derive/src/model/codegen/retrieve_impl.rs
@@ -52,7 +52,7 @@ impl ToTokens for RetrieveImpl {
                         .await
                         .map(#model::from)
                         .optional()
-                        .map_err(|e| Self::Error::from(editoast_models::model::Error::from(e)))
+                        .map_err(|e| Self::Error::from(editoast_models::Error::from(e)))
                 }
             }
         });
