+++ editoast/src/models/infra.rs
@@ -17,7 +17,6 @@ use diesel::sql_query;
 use diesel::sql_types::BigInt;
 use diesel_async::RunQueryDsl;
 use editoast_derive::Model;
-use editoast_models::model;
 use educe::Educe;
 use serde::Deserialize;
 use serde::Serialize;
@@ -102,7 +101,10 @@ impl Infra {
             .collect()
     }
 
-    pub async fn bump_version(&mut self, conn: &mut DbConnection) -> Result<(), model::Error> {
+    pub async fn bump_version(
+        &mut self,
+        conn: &mut DbConnection,
+    ) -> Result<(), editoast_models::Error> {
         self.version += 1;
         self.modified = Utc::now();
         self.save(conn).await
@@ -111,7 +113,7 @@ impl Infra {
     pub async fn bump_generated_version(
         &mut self,
         conn: &mut DbConnection,
-    ) -> Result<(), model::Error> {
+    ) -> Result<(), editoast_models::Error> {
         self.generated_version = Some(self.version);
         self.save(conn).await
     }
