+++ editoast/src/views/round_trips.rs
@@ -43,7 +43,7 @@ enum RoundTripsError {
     #[error("Database error")]
     #[editoast_error(status = 500)]
     #[from(forward)]
-    Database(editoast_models::model::Error),
+    Database(editoast_models::Error),
 }
 
 /// Represents a collection of round trips and one-way
