+++ editoast/src/views/scenario.rs
@@ -98,7 +98,7 @@ pub enum ScenarioError {
 
     #[error(transparent)]
     #[editoast_error(status = 500)]
-    Database(#[from] editoast_models::model::Error),
+    Database(#[from] editoast_models::Error),
 }
 
 #[editoast_derive::openapi_schema]
