+++ editoast/src/views/scenario/macro_nodes.rs
@@ -46,7 +46,7 @@ enum MacroNodeError {
 
     #[error(transparent)]
     #[editoast_error(status = 500)]
-    Database(#[from] editoast_models::model::Error),
+    Database(#[from] editoast_models::Error),
 }
 
 #[derive(IntoParams, Deserialize)]
