+++ editoast/src/views/stdcm_logs.rs
@@ -38,7 +38,7 @@ enum StdcmLogError {
 
     #[error(transparent)]
     #[editoast_error(status = 500)]
-    Database(#[from] editoast_models::model::Error),
+    Database(#[from] editoast_models::Error),
 }
 
 #[editoast_derive::openapi_schema]
