+++ editoast/src/views/study.rs
@@ -46,7 +46,7 @@ pub enum StudyError {
     StartDateAfterEndDate,
     #[error(transparent)]
     #[editoast_error(status = 500)]
-    Database(#[from] editoast_models::model::Error),
+    Database(#[from] editoast_models::Error),
 }
 
 #[editoast_derive::openapi_schema]
