+++ editoast/src/views/timetable/paced_train.rs
@@ -62,7 +62,7 @@ enum PacedTrainError {
     ExceptionNotFound { exception_key: String },
     #[error(transparent)]
     #[editoast_error(status = 500)]
-    Database(#[from] editoast_models::model::Error),
+    Database(#[from] editoast_models::Error),
 }
 
 #[editoast_derive::openapi_schema]
