+++ editoast/src/views/timetable/train_schedule.rs
@@ -99,7 +99,7 @@ pub enum TrainScheduleError {
     SimulationFailed { train_schedule_id: i64 },
     #[error(transparent)]
     #[editoast_error(status = 500)]
-    Database(#[from] editoast_models::model::Error),
+    Database(#[from] editoast_models::Error),
 }
 
 #[derive(IntoParams, Deserialize)]
